Separate new sphinx directive
