+++ sphinxgallery/gen_gallery.py
@@ -95,7 +95,7 @@ def setup(app):
     app.add_config_value('sphinxgallery_conf', gallery_conf, 'html')
     app.add_stylesheet('gallery.css')
 
-    app.add_directive('glr_thumb', GlrThumb)
+
     app.connect('builder-inited', generate_gallery_rst)
 
     app.connect('build-finished', embed_code_links)
